+++ src/argilla/server/alembic/versions/7cbcccf8b57a_create_metadata_properties_table.py
@@ -34,12 +34,12 @@ def upgrade() -> None:
     op.create_table(
         "metadata_properties",
         sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column("name", sa.Text(), nullable=False),
+        sa.Column("name", sa.String(), nullable=False),
+        sa.Column("title", sa.Text(), nullable=False),
         # TODO: We should move type column to settings as an attribute (as we are already doing for Fields and Questions).
         sa.Column("type", sa.Text(), nullable=False),
-        sa.Column("description", sa.Text()),
+        sa.Column("allowed_roles", sa.JSON(), server_default="[]", nullable=False),
         sa.Column("settings", sa.JSON(), nullable=False),
-        sa.Column("allowed_roles", sa.JSON, server_default="[]"),
         sa.Column("dataset_id", sa.Uuid(), nullable=False),
         sa.Column("inserted_at", sa.DateTime(), nullable=False),
         sa.Column("updated_at", sa.DateTime(), nullable=False),