+++ tests/integration/client/sdk/v1/test_datasets.py
@@ -39,7 +39,6 @@ from argilla.client.sdk.v1.datasets.api import (
     get_records,
     list_datasets,
     publish_dataset,
-    set_suggestion,
 )
 from argilla.client.sdk.v1.datasets.models import (
     FeedbackDatasetModel,
@@ -51,6 +50,7 @@ from argilla.client.sdk.v1.datasets.models import (
     FeedbackRecordsModel,
     FeedbackSuggestionModel,
 )
+from argilla.client.sdk.v1.records.api import set_suggestion
 from argilla.server.models import DatasetStatus, User, UserRole
 
 from tests.factories import (